refactor(#step2): 비속어 관련 client support(공통, common) 패키징 이동
