+++ src/main/java/kitchenpos/products/application/ProductService.java
@@ -3,8 +3,8 @@ package kitchenpos.products.application;
 import kitchenpos.products.dto.ProductChangePriceRequest;
 import kitchenpos.products.dto.ProductCreateRequest;
 import kitchenpos.products.dto.ProductDetailResponse;
-import kitchenpos.products.infra.PurgomalumClient;
 import kitchenpos.products.tobe.domain.*;
+import kitchenpos.support.infra.PurgomalumClient;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
