+++ src/main/java/kitchenpos/support/infra/DefaultPurgomalumClient.java
@@ -1,6 +1,5 @@
 package kitchenpos.support.infra;
 
-import kitchenpos.products.infra.PurgomalumClient;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
