+++ src/main/java/kitchenpos/support/infra/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.infra;
+package kitchenpos.support.infra;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
