+++ src/test/java/kitchenpos/menus/application/MenuServiceTest.java
@@ -7,10 +7,10 @@ import kitchenpos.menus.dto.MenuProductElement;
 import kitchenpos.menus.mapper.MenuMapper;
 import kitchenpos.menus.tobe.domain.*;
 import kitchenpos.products.infra.FakePurgomalumClient;
-import kitchenpos.products.infra.PurgomalumClient;
 import kitchenpos.products.tobe.domain.InMemoryProductRepository;
 import kitchenpos.products.tobe.domain.Product;
 import kitchenpos.products.tobe.domain.ProductRepository;
+import kitchenpos.support.infra.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
