+++ src/test/java/kitchenpos/products/application/ProductServiceTest.java
@@ -8,12 +8,12 @@ import kitchenpos.products.dto.ProductChangePriceRequest;
 import kitchenpos.products.dto.ProductCreateRequest;
 import kitchenpos.products.dto.ProductDetailResponse;
 import kitchenpos.products.infra.FakePurgomalumClient;
-import kitchenpos.products.infra.PurgomalumClient;
 import kitchenpos.products.tobe.domain.InMemoryProductRepository;
 import kitchenpos.products.tobe.domain.Product;
 import kitchenpos.products.tobe.domain.ProductDomainService;
 import kitchenpos.products.tobe.domain.ProductRepository;
 import kitchenpos.support.event.ProductPriceChangedEvent;
+import kitchenpos.support.infra.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
